remove unnecessary readinessProbes
* readinessProbes all combine together to form
PodStatus, which when any readinessProbe fails means
that the Pod will not receive traffic if it is connected
to a Service. I guess it doesn't matter now since no
Service is connecting to HAProxy-Router, but if we do
that later, we will be supported.
* update livenessProbes to better allow for really, really
slow startups. We don't care if Logstash takes forever to
start, as long as in the long haul it can keep up with HAProxy
and the amount of logs it generates, so we can wait 600s
before starting livenessProbes